requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.8.1
clikit==0.6.2
cloudpickle==2.0.0
crashtest==0.3.1
cryptography==37.0.2
decorator==5.1.1
distlib==0.3.4
Django==4.0.5
django-extensions==3.1.5
django-simple-history==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.6.1
executing==0.8.3
filelock==3.7.1
html5lib==1.1
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
ipython==8.4.0
jedi==0.18.1
jeepney==0.8.0
jsonschema==4.6.0
keyring==23.6.0
lockfile==0.12.2
matplotlib-inline==0.1.3
msgpack==1.0.4
numpy==1.22.4
packaging==20.9
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.8.2
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.8.0
ptyprocess==0.7.0
pudb==2022.1.1
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.14.1
Pyfhel==3.1.4
Pygments==2.12.0
PyJWT==2.4.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
scipy==1.7.2
SecretStorage==3.3.2
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomlkit==0.11.0
tornado==6.1
traitlets==5.2.2.post1
uritemplate==4.1.1
urllib3==1.26.9
urwid==2.1.2
urwid-readline==0.13
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
xgboost==1.5.0
zipp==3.8.0
